refactor: removed puris.apikey env variable, updated env files
