+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/stock/domain/model/ItemStock.java
@@ -28,7 +28,7 @@ import lombok.experimental.SuperBuilder;
 import org.eclipse.tractusx.puris.backend.common.api.logic.service.PatternStore;
 import org.eclipse.tractusx.puris.backend.masterdata.domain.model.Material;
 import org.eclipse.tractusx.puris.backend.masterdata.domain.model.Partner;
-import org.eclipse.tractusx.puris.backend.stock.logic.dto.itemstocksamm.ItemUnitEnumeration;
+import org.eclipse.tractusx.puris.backend.stock.domain.model.measurement.ItemUnitEnumeration;
 
 import java.util.Date;
 import java.util.UUID;
