+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/stock/domain/model/measurement/ItemUnitEnumeration.java
@@ -17,7 +17,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.eclipse.tractusx.puris.backend.stock.logic.dto.itemstocksamm;
+package org.eclipse.tractusx.puris.backend.stock.domain.model.measurement;
 
 import com.fasterxml.jackson.annotation.JsonValue;
