+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/stock/logic/dto/StockDto.java
@@ -31,7 +31,7 @@ import lombok.ToString;
 import org.eclipse.tractusx.puris.backend.common.api.logic.service.PatternStore;
 import org.eclipse.tractusx.puris.backend.masterdata.logic.dto.MaterialDto;
 import org.eclipse.tractusx.puris.backend.masterdata.logic.dto.PartnerDto;
-import org.eclipse.tractusx.puris.backend.stock.logic.dto.itemstocksamm.ItemUnitEnumeration;
+import org.eclipse.tractusx.puris.backend.stock.domain.model.measurement.ItemUnitEnumeration;
 
 import java.io.Serializable;
 import java.util.Date;
