+++ backend/src/test/resources/application.properties
@@ -1,7 +1,6 @@
 # Server Config
 server.port=${SERVER_PORT:8081}
 puris.demonstrator.role=${PURIS_DEMONSTRATOR_ROLE:customer}
-puris.apiversion=${PURIS_APIVERSION:1.0.0}
 puris.edr.endpoint=${PURIS_EDR_ENDPOINT:http://customer-backend:8081/catena/edrendpoint}
 puris.edr.deletiontimer=${PURIS_EDR_DELETIONTIMER:2}
 puris.request.serverendpoint=${PURIS_REQUEST_SERVERENDPOINT:http://customer-backend:8081/catena/item-stock/request}
