+++ charts/puris/templates/backend-deployment.yaml
@@ -83,8 +83,6 @@ spec:
               value: {{ index .Values.backend "puris" "jpa" "hibernate" "ddl-auto" }} #Need to do workaround because of '-' in ddl-auto
             - name: JPA_PROPERTIES_HIBERNATE_ENABLE.LAZY.LOAD.NO.TRANS
               value: "{{ .Values.backend.puris.jpa.properties.hibernate.enable_lazy_load_no_trans }}"
-            - name: PURIS_APIVERSION
-              value: "{{ .Values.backend.puris.api.version }}"
             - name: PURIS_API_KEY
               value: "{{ .Values.backend.puris.api.key}}"
             - name: PURIS_DEMONSTRATOR_ROLE
