+++ local/tractus-x-edc/config/customer/puris-backend.properties
@@ -1,6 +1,5 @@
 server.port=8081
 puris.demonstrator.role=customer
-puris.apiversion=1.0.0
 puris.edr.endpoint=http://customer-backend:8081/catena/edrendpoint
 puris.edr.deletiontimer=2
 puris.request.serverendpoint=http://customer-backend:8081/catena/item-stock/request
