+++ local/tractus-x-edc/config/supplier/puris-backend.properties
@@ -1,6 +1,5 @@
 server.port=8082
 puris.demonstrator.role=supplier
-puris.apiversion=1.0.0
 puris.edr.endpoint=http://supplier-backend:8082/catena/edrendpoint
 puris.edr.deletiontimer=2
 puris.request.serverendpoint=http://supplier-backend:8082/catena/item-stock/request
